replaced develop version of subsample_bioseqfile.py with one from subsample-fastq-38 branch
